backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase>=2.12.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
jose==1.0.0
pytest==7.4.4
httpx==0.27.2
music21==9.1.0
pretty_midi==0.2.10
openai==1.12.0
numpy==1.26.4
scipy==1.12.0
mido==1.3.0
pyfluidsynth==1.3.2
pydub==0.25.1
requests==2.31.0
setuptools
